--- tests/distributed/conftest.py
@@ -1,43 +1,0 @@
-# coding=utf-8
-# Copyright 2024 The HuggingFace Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import pytest
-
-
-# This hook is run before the default pytest_runtest_call
-@pytest.hookimpl(tryfirst=True)
-def pytest_runtest_call(item):
-    # We want to use our own launching function for distributed tests
-    if getattr(item.cls, "is_dist_test", False):
-        dist_test_class = item.cls()
-        dist_test_class(item._request)
-        item.runtest = lambda: True  # Dummy function so test is not run twice
-
-
-# We allow DistributedTest to reuse distributed environments. When the last
-# test for a class is run, we want to make sure those distributed environments
-# are destroyed.
-def pytest_runtest_teardown(item, nextitem):
-    if getattr(item.cls, "reuse_dist_env", False) and not nextitem:
-        dist_test_class = item.cls()
-        for num_procs, pool in dist_test_class._pool_cache.items():
-            dist_test_class._close_pool(pool, num_procs, force=True)
-
-
-@pytest.hookimpl(tryfirst=True)
-def pytest_fixture_setup(fixturedef, request):
-    if getattr(fixturedef.func, "is_dist_fixture", False):
-        dist_fixture_class = fixturedef.func()
-        dist_fixture_class(request)
