+++ tests/utils.py
@@ -14,28 +14,35 @@
 # limitations under the License.
 """Various utilities used in multiple tests."""
 
+import contextlib
+import functools
+import inspect
 import os
 import random
 import string
-from typing import Dict, Optional, Set, Tuple
+from typing import Callable, Dict, List, Optional, Tuple, Type
 
 import torch
 from datasets import Dataset, DatasetDict
 from huggingface_hub import CommitOperationDelete, HfApi, create_repo, delete_repo, get_token, login, logout
 from huggingface_hub.utils import RepositoryNotFoundError
-from transformers import PretrainedConfig, PreTrainedModel
+from transformers import AutoConfig, PreTrainedModel
 from transformers.testing_utils import ENDPOINT_STAGING
 
+from optimum.neuron.distributed import lazy_load_for_parallelism
 from optimum.neuron.utils.cache_utils import (
     delete_custom_cache_repo_name_from_hf_home,
     load_custom_cache_repo_name_from_hf_home,
     set_custom_cache_repo_name_in_hf_home,
 )
+from optimum.neuron.utils.patching import DynamicPatch, Patcher
 from optimum.utils import logging
 from optimum.utils.testing_utils import TOKEN, USER
 
 
 logger = logging.get_logger(__name__)
+
+SEED = 42
 
 
 def get_random_string(length) -> str:
@@ -43,7 +50,8 @@ def get_random_string(length) -> str:
     return "".join(random.choice(letters) for _ in range(length))
 
 
-def create_dummy_dataset(input_specs: Dict[str, Tuple[int, ...]], num_examples: int) -> Dataset:
+def create_dummy_dataset(input_specs: Dict[str, Tuple[Tuple[int, ...], torch.dtype]], num_examples: int) -> Dataset:
+
     def gen():
         for _ in range(num_examples):
             yield {name: torch.rand(shape) for name, shape in input_specs.items()}
@@ -75,52 +83,126 @@ def create_dummy_text_classification_dataset(
     return ds
 
 
-class MyTinyModel(PreTrainedModel):
-    config_class = PretrainedConfig
-
-    def __init__(self, config: PretrainedConfig):
-        super().__init__(config)
-        self.linears = torch.nn.ModuleList([torch.nn.Linear(1, 1) for _ in range(config.num_linears)])
-        self.relu = torch.nn.ReLU()
-        self.criterion = torch.nn.MSELoss()
-
-    def forward(self, x, labels=None):
-        for lin in self.linears:
-            x = lin(x)
-            x = self.relu(x)
-        if labels is not None:
-            loss = self.criterion(x, labels)
-            outputs = (loss, x)
-        else:
-            outputs = (x,)
-        return outputs
+def generate_input_ids(vocab_size: int, batch_size: int, sequence_length: int) -> torch.Tensor:
+    return torch.randint(0, vocab_size, (batch_size, sequence_length))
 
 
-def create_tiny_pretrained_model(
-    num_linears: int = 1,
-    random_num_linears: bool = False,
-    max_num_linears: int = 20,
-    visited_num_linears: Optional[Set[int]] = None,
-) -> PreTrainedModel:
-    if visited_num_linears is not None:
-        if len(visited_num_linears) == max_num_linears:
-            raise RuntimeError(
-                f"There are too many tests for the maximum number of linears allowed ({max_num_linears}), please "
-                "increase it."
-            )
+def create_dummy_causal_lm_dataset(
+    vocab_size: int,
+    num_train_examples: int,
+    num_eval_examples: int,
+    num_test_examples: Optional[int] = None,
+) -> DatasetDict:
+    if num_test_examples is None:
+        num_test_examples = num_eval_examples
+
+    def create_gen(num_examples):
+        def gen():
+            for _ in range(num_examples):
+                input_ids = generate_input_ids(vocab_size, 1, 32)
+                yield {
+                    "input_ids": input_ids,
+                    "labels": input_ids,
+                }
+
+        return gen
+
+    ds = DatasetDict()
+    ds["train"] = Dataset.from_generator(create_gen(num_train_examples))
+    ds["eval"] = Dataset.from_generator(create_gen(num_eval_examples))
+    ds["test"] = Dataset.from_generator(create_gen(num_test_examples))
+
+    return ds
+
+
+def default_data_collator_for_causal_lm(features: List[Dict[str, torch.Tensor]]) -> Dict[str, torch.Tensor]:
+    feature_names = features[0].keys()
+    return {k: torch.cat([torch.tensor(feature[k]) for feature in features], dim=0) for k in feature_names}
+
+
+def static_initializer_seed(initialization_function: Callable, seed: int):
+    @functools.wraps(initialization_function)
+    def wrapper(*args, **kwargs):
+        from transformers import set_seed
+
+        set_seed(seed)
+        return initialization_function(*args, **kwargs)
+
+    return wrapper
+
+
+@contextlib.contextmanager
+def create_static_seed_patcher(model_class: Type["PreTrainedModel"], seed: int):
+    """
+    Context manager that resets the seed to a given value for every initialization function.
+    This is useful because lazy initialization works but does not respect the random state of the non-lazy case.
+    This allows us to test that lazy initialization works if we ignore the random seed.
+    """
+    specialized_static_initializer_seed = functools.partial(static_initializer_seed, seed=seed)
+
+    inspect.getmodule(model_class).__name__
+    dynamic_patch = DynamicPatch(specialized_static_initializer_seed)
+    patcher = Patcher(
+        [
+            # (fully_qualified_method_name, dynamic_patch),
+            ("torch.nn.Embedding.reset_parameters", dynamic_patch),
+            ("torch.nn.Linear.reset_parameters", dynamic_patch),
+            ("torch.Tensor.normal_", dynamic_patch),
+            ("neuronx_distributed.parallel_layers.layers.ColumnParallelLinear.init_weight_cpu", dynamic_patch),
+            ("neuronx_distributed.parallel_layers.layers.RowParallelLinear.init_weight_cpu", dynamic_patch),
+            (
+                "neuronx_distributed.modules.qkv_linear.GQAQKVColumnParallelLinear._init_per_layer_weight",
+                dynamic_patch,
+            ),
+            ("neuronx_distributed.modules.qkv_linear.GQAQKVColumnParallelLinear._init_per_layer_bias", dynamic_patch),
+        ]
+    )
+    with patcher:
+        try:
+            yield
+        finally:
+            pass
+
+
+def get_model(
+    model_class: Type["PreTrainedModel"],
+    model_name_or_path: str,
+    tp_size: int = 1,
+    pp_size: int = 1,
+    lazy_load: bool = False,
+    from_config: bool = False,
+    use_static_seed_patcher: bool = False,
+    add_random_noise: bool = False,
+    config_overwrite: Optional[Dict[str, str]] = None,
+) -> "PreTrainedModel":
+    if lazy_load:
+        ctx = lazy_load_for_parallelism(tensor_parallel_size=tp_size, pipeline_parallel_size=pp_size)
     else:
-        visited_num_linears = set()
+        ctx = contextlib.nullcontext()
+    if use_static_seed_patcher:
+        seed_patcher = create_static_seed_patcher(model_class, SEED)
+    else:
+        seed_patcher = contextlib.nullcontext()
+    with ctx:
+        with seed_patcher:
+            config = AutoConfig.from_pretrained(model_name_or_path)
+            if config_overwrite is not None:
+                for key, value in config_overwrite.items():
+                    attr_type = type(getattr(config, key))
+                    setattr(config, key, attr_type(value))
+            if from_config:
+                model = model_class(config)
+            else:
+                model = model_class.from_pretrained(model_name_or_path, config=config, ignore_mismatched_sizes=True)
 
-    if random_num_linears:
-        num_linears = random.randint(1, max_num_linears)
-        while num_linears in visited_num_linears:
-            num_linears = random.randint(1, max_num_linears)
-        visited_num_linears.add(num_linears)
+    if getattr(model.config, "problem_type", None) is None:
+        model.config.problem_type = "single_label_classification"
 
-    config = PretrainedConfig()
-    config.num_linears = num_linears
+    if add_random_noise:
+        for param in model.parameters():
+            param.data.add_(torch.randn_like(param))
 
-    return MyTinyModel(config)
+    return model
 
 
 class TrainiumTestMixin:
@@ -210,17 +292,3 @@ class StagingTestMixin:
         login(TOKEN)
         self.remove_all_files_in_repo(self.CUSTOM_CACHE_REPO)
         self.remove_all_files_in_repo(self.CUSTOM_PRIVATE_CACHE_REPO)
-
-    def create_tiny_pretrained_model(self, num_linears: int = 1, random_num_linears: bool = False):
-        return create_tiny_pretrained_model(
-            num_linears=num_linears,
-            random_num_linears=random_num_linears,
-            visited_num_linears=self.visited_num_linears,
-        )
-
-    def create_and_run_tiny_pretrained_model(self, num_linears: int = 1, random_num_linears: bool = False):
-        tiny_model = self.create_tiny_pretrained_model(num_linears=num_linears, random_num_linears=random_num_linears)
-        tiny_model = tiny_model.to("xla")
-        random_input = torch.rand(1, device="xla")
-        print(tiny_model(random_input))
-        return tiny_model
